Move all functions related to reporter to separate file.
Apply reporter in one place - on check task creation.
